telegram_bot/telegram_bot/requirements.txt
aiogram==3.0.0
aiohttp
beautifulsoup4
python-dotenv
